dashboard/requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
cchardet==2.1.1
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==2.0
Cython==0.26
dnspython==1.15.0
dnsq==1.1.6
enum34==1.1.6
expiringdict==1.1.3
flanker==0.4.38
Flask==0.12.2
Flask-WTF==0.14.2
google-api-python-client==1.6.2
GoogleAppEngineCloudStorageClient==1.9.22.1
httplib2==0.10.3
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
oauth2client==4.1.2
packaging==16.8
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.18
pyparsing==2.2.0
redis==2.10.5
regex==2017.7.11
requests==2.18.1
requests-toolbelt==0.8.0
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1
WebOb==1.7.3
Werkzeug==0.12.2
WTForms==2.1
